# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Job Hunting API
  version: 1.0.0
  description: API documentation for Job Hunting app
servers:
  - url: http://localhost:4002
    description: Local server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []

# ------------------------
# AUTH ROUTES
# ------------------------
paths:
...
  /auth/reset-password/{token}:
    post:
      summary: Reset password
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                newPassword:
                  type: string
      responses:
        '200':
          description: Password reset successfully

# ------------------------
# USER ROUTES
# ------------------------
...